Make message REST action on campaigns.
