+++ app/routes/index.js
@@ -29,4 +29,3 @@ router.use('/v1/campaigns', require('./campaigns'));
 router.use('/v1/chatbot', require('./chatbot'));
 router.use('/v1/donorschoosebot', require('./donorschoosebot'));
 router.use('/v1/signups', require('./signups'));
-router.use('/v1/notifications', require('./notifications'));
